requirements.txt
allure-pytest==2.9.45
pytest==7.1.1
pytest-xdist==2.5.0
requests==2.27.1
playwright==1.32.1

